backend/requirements.txt
flask
flask-cors
gensim
dataclasses-json
panda
numpy
operator
